requirements.txt
beautifulsoup4>=4.10
praw>=7.7.1
selenium>=4.6.0
python-dotenv>=1.0.1
